api_service/pyproject.toml
[project]
name = "task-planner-bot"
version = "0.8.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.40,<2.1.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "pydantic (>=2.11.1,<2.12.0)",
    "pydantic-settings (>=2.9.1,<2.10.0)",
    "pytest (>=8.3.5,<8.4.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "alembic (>=1.15.2,<1.16.0)",
    "pyjwt (>=2.10.1,<2.11.0)",
    "passlib (>=1.7.4,<1.8.0)",
    "bcrypt (>=4.3.0,<4.4.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "tabulate (>=0.9.0,<0.10.0)",
    "redis (>=6.2.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

